alpha_vantage/requirements.txt
python-dotenv>=1.2.2
alpha_vantage>=3.0.0
pandas